Now we are using System.loadLibrary and System.load properly
Because in Java 15 it seems the hack to reload the classpath at runtime
will not longer work. Specifically the usr_paths is no longer a private
member. Using System.load() as oppose to System.loadLibrary() gets
around the need to munge the classpath at runtime and thus makes this
work with Java 17 just fine. We left the System.loadLibrary path
specifically for CI as it's not packaged in a jar at that point and the
native library file is not found to be extracted